feat: Create a service provider's calendar layout
Add a query "findAllByProviderIdFromDate" in the 'TermRepository" that selects all
terms shered by provider after a specific date. Add a method "findAllFutureByProviderId"
in the "TermService" that returns all future terms of a specific provider. Refactor the
enum Status, so that it has a constructor. Add a method "getProviderCalendar" in the
"ServiceProviderController" which adds a query result to the model and returns the "providers/calendar".
Create a calendar layout in the file "calendar.html". Update the "messages.properties" with new texts.