feat: Add terms filtering in the provider archive by a client name
Refactor the "archive.html", so that it has a form with a client name input and the submit button.
Add a hidden input with a providerId. Add a script in the "layout.html" with a function that
sets a providerId value from the local storage. Modify methods in the "TermService" which filter
using clientId, so that they filter using client name. Add a query in the "TermRepository" that
selects clients by a specific name. Create a method "processFindForm" in the "ServiceProviderController"
that adds all past terms of a specific provider appointed by a selected client to the model.