feat: Apply created authentication to clients and service providers
Make the entity class "Person" implements the "UserDetails" interface. Create
a class "UserAuthenticationService" that returns an authenticated client or
a service provider based on a given email. Add queries in the "ClientRepository"
and the "ServiceProviderRepository" that select a client and a provider having
a specified username. Add a method "getAuthenitacedProviderId" that returns an
id number of an authenticated provider in the ProviderController to let load
personalised content.