feat: Create find terms layout
Create a class "ServiceProviderTerm" that combines a service provider with the
closest term. Create a file "find.html" with a table and inputs for filtering
results. Add a method "findTerms" in the "ClientController" that add selected
instances of "ServiceProviderTerm" to the model. Replace methods in the
"ServiceProviderService" which find providers compatible with one condition
with a method "findAllByNameAndCityAndFieldFromDateInTermOrder". Add a query in
the "TermRepository" that selects all terms compatible with method's conditions.