feat: Add modals with terms' details in the client's terms
Create a modal for every table row in the file "terms.html". Add a "cancel
button" in modals for future terms. Create a method "cancelTerm" in the
"TermController" that calls a method canceling term with a specified id.
Add a transactional method "cancelById" in the "TermService" which call an
update in the database. Create a query in the "TermRepository" that makes
a termRealizedStatus free.