feat: Add saving a new service provider image
Create a modal in the "account.html" that let load a new image from
computer's files and shows an image preview. Add a method "addProviderImage"
in the "ServiceProviderController" that provokes saving an image. Create
a class "ImageService" and its method "saveProviderImage" that parses an
image into a byte array. Add a query in the "ServiceProviderRepository" that
update the provider of a given id by adding a new image.