feat: Add booking a term by a client
Add a href path in the "provider.html". Create a method "bookTerm" in the
"ClientController" which calls the booking service and redirects to client's
terms. Add a method "bookTermByClientId" in the "TermService" that books
term for the given client and makes term's status full. Create queries in
the "TermRepository" that update status and clientId of the given term.