feat: Add possibility of adding terms by a provider
Create a button in the "calendar.html" that opens a modal containing inputs that let provider
create a sequence of terms. Add a method "addNewTerms" in the "ProviderController" which
validates arguments and calls a method saving terms. Create a method in the "TermService"
that creates and saves a proper number of terms based on the given parameters.