feat: Add providers' photos in the find provider layout
Create a new column in the "find.html" with providers' images. Modify the
"ServiceProviderTerm", so that it contains String with provider's image. Add
creating image's String from bytes table in the "ServiceProviderController",
so that it could be added to mapped provider in the method returning a list
of ServiceProviderTerms's objects.