feat: Add updating client's personal details
Add a button in the "account.html" which opens a file "details.html".
Create a file "details.html" with a form that let update client's details.
Add a method "getUpdatingDetailsForm" that returns the "details.html" and
a method "updateAccountDetails" which calls a method updating client's
account and redirects page to the client's account. Create a method in the
"ClientService" that saves client's account settings.