feat: Add updating service provider's personal details
Add a button in the "account.html" which opens a file "details.html". Create
a file "details.html" with a form that let update provider's details. Delete
the modal in the "account.html" that gave a possibility to change a provider
image. Add a method "getUpdatingDetailsForm" that returns the "details.html"
and a method "updateAccountDetails" which calls a method updating provider's
account and redirects page to the provider's account. Create a method in the
"ServiceProviderService" that saves provider's account settings.