feat: Add user registration
Create a file "signup.html" with a form that let create a new user. Add a method
"showRegistrationForm" that returns the "signup" page and a method "registerUserAccount"
that handles exceptions and calls a method registering user. Create a class "UserDto". Add
a method "registerNewUserAccount" in the "UserAuthenticationService" that creates a new
account. Create package "exceptions" with an exception "UserAlreadyExistException".